Put dump-metadata code in separate file
